New: Extract backend releated JS to own files
