+++ Resources/Private/JavaScripts/Helper/Backend/index.js
@@ -1,0 +1,9 @@
+import applyBackendAdjustments from './applyBackendAdjustments';
+import removeEmptySlides from './removeEmptySlides';
+import selectSlideForNode from './selectSlideForNode';
+
+export {
+    applyBackendAdjustments,
+    removeEmptySlides,
+    selectSlideForNode
+};
