+++ Resources/Private/JavaScripts/index.js
@@ -1,13 +1,7 @@
 import $ from 'jquery';
 import 'slick-carousel';
-import {
-    getCollectionValueByPath,
-    getSliderPositionByIndex,
-    getItemByKeyValue,
-    isNil
-} from './Helper';
-
-const sliderElements = [];
+import { getCollectionValueByPath, isNil } from './Helper';
+import { applyBackendAdjustments } from './Helper/Backend';
 
 const adjustDimensions = slick => {
     const width = parseInt(slick.listWidth);
@@ -20,64 +14,6 @@ const adjustDimensions = slick => {
     });
 };
 
-const removeEmptySlides = slick => {
-    // use reverse to prevent index changes when removing slides
-    const slides = Array.from(slick.$slides).reverse();
-    slides.forEach(slide => {
-        const hasInnerSlide = slide.querySelector('.slide__inner');
-        if (isNil(hasInnerSlide)) {
-            const slickIndex = slide.getAttribute('data-slick-index');
-            slick.slickRemove(slickIndex);
-        }
-    });
-};
-
-const collectSlideMetaInformation = () => {
-    const slides = document.querySelectorAll('.slick-slide:not(.slick-cloned)');
-    const slideData = [];
-    slides.forEach(slide => {
-        const innerSlide = slide.querySelector('.slide__inner');
-        const parentSlider = $(slide).closest('.slick-slider');
-        if (!isNil(innerSlide) && !isNil(parentSlider)) {
-            slideData.push({
-                index: slide.getAttribute('data-slick-index'),
-                identifier: innerSlide.getAttribute('data-slide-identifier'),
-                parentIdentifier: parentSlider.data('slider-identifier')
-            });
-        }
-    });
-    return slideData;
-};
-
-const selectSlideForNode = node => {
-    const slideMetaInformation = collectSlideMetaInformation();
-    const selectSlide = getItemByKeyValue(
-        slideMetaInformation,
-        'identifier',
-        node.name
-    );
-    if (!isNil(selectSlide)) {
-        const slider = getItemByKeyValue(
-            sliderElements,
-            'identifier',
-            selectSlide.parentIdentifier
-        );
-        const sliderPosition = getSliderPositionByIndex(
-            Array.from(slider.slider.$slides),
-            selectSlide.index
-        );
-
-        if (sliderPosition >= 0) {
-            slider.slider.slickGoTo(sliderPosition);
-
-            // resize window to rerender the backend ui controls
-            const resizeEvent = window.document.createEvent('UIEvents');
-            resizeEvent.initUIEvent('resize', true, false, window, 0);
-            window.dispatchEvent(resizeEvent);
-        }
-    }
-};
-
 const sliders = $('[data-slick]');
 sliders.on('setPosition', function(event, slick) {
     adjustDimensions(slick);
@@ -85,15 +21,7 @@ sliders.on('setPosition', function(event, slick) {
 
 sliders.on('init', function(event, slick) {
     if ('neos' in window) {
-        const slider = slick.$slider.get(0);
-        if (!isNil(slider)) {
-            const identifier = slider.getAttribute('data-slider-identifier');
-            sliderElements.push({ identifier: identifier, slider: slick });
-        }
-
-        // For some reason we have in the neos backend for each slide one additional
-        // empty slide. This is bit hacky way get rid of them.
-        removeEmptySlides(slick);
+        applyBackendAdjustments(slick);
     }
     adjustDimensions(slick);
 });
@@ -108,13 +36,3 @@ Array.from(sliders).forEach(slider => {
         configuration != '' ? JSON.parse(configuration) : {};
     $(slider).slick(configurationObject);
 });
-
-// list for the event Neos.NodeSelected in the neos backend to select a slide
-if ('neos' in window) {
-    document.addEventListener('Neos.NodeSelected', function(event) {
-        const node = getCollectionValueByPath(event, 'detail.node');
-        if (!isNil(node)) {
-            selectSlideForNode(node);
-        }
-    });
-}
